requirements.txt
clip==1.0
matplotlib==3.7.2
numpy==1.25.2
ruptures==1.1.9
Pillow==10.3.0
scipy==1.14.0
seaborn==0.13.2
torch==2.0.1+cu117
nltk==3.8.1
pandas==2.0.3
sentence-transformers==2.6.1
opencv-python==4.10.0.84
scikit-learn==1.3.0